Move tests comparing relation members with response to integration tests
